+++ jni/include/simd/similarity_function/similarity_function.h
@@ -14,7 +14,11 @@ namespace knn_jni::simd::similarity_function {
         // L2 for FP16
         FP16_L2,
         SQ_IP,
-        SQ_L2
+        SQ_L2,
+        // Max inner product for BF16.
+        BF16_MAXIMUM_INNER_PRODUCT,
+        // L2 for BF16
+        BF16_L2
     };
 
     struct SimilarityFunction;