+++ src/main/java/org/opensearch/knn/index/codec/KNN1040Codec/KNN1040ScalarQuantizedVectorScorer.java
@@ -243,8 +243,9 @@ public class KNN1040ScalarQuantizedVectorScorer extends Lucene104ScalarQuantized
                 targetCorrectiveTerms.additionalCorrection(),
                 targetCorrectiveTerms.quantizedComponentSum(),
                 addressAndSize,
-                similarityFunction == VectorSimilarityFunction.MAXIMUM_INNER_PRODUCT
+                (similarityFunction == VectorSimilarityFunction.MAXIMUM_INNER_PRODUCT
                     || similarityFunction == VectorSimilarityFunction.COSINE
+                    || similarityFunction == VectorSimilarityFunction.DOT_PRODUCT)
                         ? SimdVectorComputeService.SimilarityFunctionType.SQ_IP.ordinal()
                         : SimdVectorComputeService.SimilarityFunctionType.SQ_L2.ordinal(),
                 dimension,